Create sphere_setup_xyz.py
Right now, the metric and shift functions for performing soft sphere simulations using Jax MD. So, this lets one do NVT, NVE, FIRE... however, there's an as of yet unsolved bug: I expect the shift to work as r+dr, where dr is the shift projected into the tangent space of the sphere, but instead it appears only to work when the shift is written r-dr. I think this has something to do with the correctness of the projection formalism.